add spring security

+++ SaleWebV1/src/main/webapp/resources/js/cart.js
@@ -17,7 +17,10 @@ function addToCart(endpoint, id, name, price) {
         }
     }).then(res => res.json()).then(data => {
         console.info(data)
-    })
+        let counters = document.getElementsByClassName("cart-counter");
+        for (let d of counters)
+            d.innerText = data.totalQuantity;
+    });
 }
 
 function updateItem(endpoint, obj) {
@@ -30,14 +33,47 @@ function updateItem(endpoint, obj) {
             "Content-Type": "application/json"
         }
     }).then(res => res.json()).then(data => {
-        console.info(data)
-    })
+        let counters = document.getElementsByClassName("cart-counter");
+        for (let d of counters)
+            d.innerText = data.totalQuantity;
+        
+        let amounts = document.getElementsByClassName("cart-amount");
+        for (let d of amounts)
+            d.innerText = parseFloat(data.totalAmount).toLocaleString("en-US");
+    });
 }
 
-function deleteItem(endpoint) {
+function deleteItem(endpoint, id) {
+    if (confirm("Bạn chắc chắn xóa không?") === true) {
+        fetch(endpoint, {
+            method: "delete"
+        }).then(res => res.json()).then(data => {
+            let el = document.getElementById(`cart${id}`);
+            el.style.display = "none";
+
+            let counters = document.getElementsByClassName("cart-counter");
+            for (let d of counters)
+                d.innerText = data.totalQuantity;
+            
+            let amounts = document.getElementsByClassName("cart-amount");
+        for (let d of amounts)
+            d.innerText = parseFloat(data.totalAmount).toLocaleString("en-US");
+        });
+    }
+}
+
+function pay(endpoint) {
     fetch(endpoint, {
-        method: "delete"
-    }).then(res => res.json()).then(data => {
-        console.info(data);
-    });
+        method: "post"
+    }).then(res => {
+        console.info(res);
+        if (res.status === 200) {
+            let e = document.getElementById("content");
+            e.innerText = "Đơn hàng đã được ghi nhận";
+            
+            let counters = document.getElementsByClassName("cart-counter");
+            for (let d of counters)
+                d.innerText = 0;
+        } 
+    })
 }
